refactor(comment-query):optimize the comment-query with middleware of redis cache

+++ oursphere-comment/oursphere-comment-biz/src/main/java/com/theoyu/oursphere/comment/biz/model/entity/CommentPO.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -45,7 +44,7 @@ public class CommentPO {
 
     private Long childCommentTotal;
 
-    private BigDecimal heat;
+    private Double heat;
 
     private Long firstReplyCommentId;
